Partially completed conversion to new prompt system
