+++ paperqa/prompts.py
@@ -1,0 +1,71 @@
+from datetime import datetime
+
+from langchain.prompts import PromptTemplate
+
+summary_prompt = PromptTemplate(
+    input_variables=["answer", "text"],
+    template="Summarize the text below to help answer a question. "
+    "Do not directly answer the question, instead summarize "
+    "to give evidence to help answer the question. Include direct quotes. "
+    'Reply "Not applicable" if text is irrelevant. '
+    "Use {summary_length}. At the end of your response, provide a score from 1-10 on a newline "
+    "indicating relevance to question. Do not explain your score. "
+    "\n\n"
+    "{text}\n"
+    "Extracted from {text.doc.citation}\n"
+    "Question: {answer.question}\n"
+    "Relevant Information Summary:",
+)
+
+qa_prompt = PromptTemplate(
+    input_variables=["answer"],
+    template="Write an answer ({answer.answer_length}) "
+    "for the question below based on the provided context. "
+    "If the context provides insufficient information, "
+    'reply "I cannot answer". '
+    "For each part of your answer, indicate which sources most support it "
+    "via valid citation markers at the end of sentences, like (Example2012). "
+    "Answer in an unbiased, comp rehensive, and scholarly tone. "
+    "If the question is subjective, provide an opinionated answer in the concluding 1-2 sentences. \n\n"
+    "{answer.context}\n"
+    "Question: {answer.question}\n"
+    "Answer: ",
+)
+
+
+search_prompt = PromptTemplate(
+    input_variables=["question"],
+    template="We want to answer the following question: {question} \n"
+    "Provide three keyword searches (one search per line) "
+    "that will find papers to help answer the question. Do not use boolean operators. "
+    "Provide some broad and some specific searches. "
+    "Recent years are 2021, 2022, 2023.\n\n"
+    "1.",
+)
+
+
+select_paper_prompt = PromptTemplate(
+    input_variables=["question", "papers"],
+    template="Select papers to help answer the question below. "
+    "Papers are listed as $KEY: $PAPER_INFO. "
+    "Return a list of keys, separated by commas. "
+    'Return "None", if no papers are applicable. '
+    "Choose papers that are relevant, from reputable sources, and timely. \n\n"
+    "Question: {question}\n\n"
+    "{papers}\n\n"
+    "Selected keys:",
+)
+
+
+def _get_datetime():
+    now = datetime.now()
+    return now.strftime("%m/%d/%Y")
+
+
+citation_prompt = PromptTemplate(
+    input_variables=["text"],
+    template="Provide the citation for the following text in MLA Format. Today's date is {date}\n"
+    "{text}\n\n"
+    "Citation:",
+    partial_variables={"date": _get_datetime},
+)
